backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas
numpy==1.26.2
scikit-learn==1.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
joblib==1.3.2